--- src/main/java/com/example/KitchenProject/CookBook.java
@@ -1,15 +1,0 @@
-package com.example.KitchenProject;
-
-import java.util.List;
-
-public class CookBook {
-    private List<Recipe> recipes;
-
-    public List<Recipe> getRecipes() {
-        return recipes;
-    }
-
-    public void addRecipe (Recipe newRecipe){
-        recipes.add(newRecipe);
-    }
-}
