--- src/main/java/com/example/KitchenProject/CookingLevel.java
@@ -1,8 +1,0 @@
-package com.example.KitchenProject;
-
-public enum CookingLevel {
-    EASY,
-    MEDIUM,
-    HARD,
-    MASTER
-}
