+++ src/main/java/com/example/kitchenproject/HelloWorldController.java
@@ -1,4 +1,4 @@
-package com.example.KitchenProject;
+package com.example.kitchenproject;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
