+++ src/main/java/com/example/kitchenproject/KitchenProjectApplication.java
@@ -1,4 +1,4 @@
-package com.example.KitchenProject;
+package com.example.kitchenproject;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
